Adjust VBATT_START/PAUSE_CHARGING
Behavior is a bit different than in 0.7.0, and bearing in mind how this will affect the Comma Two, I did some testing in my Prius Prime.  Currently, this change allows the EON to stay for about a minute (or more, depending on the 12V battery charge) after I turn off the car (assuming it detects a Panda).  Want to keep the EON on indefinitely?  Just disconnect it from the Panda, or ensure the Panda has no power.  I'll surely need to tweak this later, but this gets the ball rolling.  Perhaps it would be wise for those w/ my LVD to lower the cutoff ~.2V or so.  Comma Two may behave differently than EON (eg. may not do a controlled poweroff).